All Test Infrastructure
Previously the gaming server just had the required infrastructure for a VPC and a subnet. Now I added an internet gateway, a route table, routes, a security group, a key pair and the actual EC2 instance that's meant to run the server. Also included a "datasources" file for the image that will be used by AWS for the OS of the EC2 instance, a hosts template that will write the IP of the generated server into a hosts file in order to provision the server and some variables that include my personal IP and other future variables that I want to include.