Elastic IP and Domain Name Mapping
Assigned an elastic IP address for the gaming server, which basically gives a static IP to an instance even when the instance is stopped or down. After that, I got the domain name "noobsquad.xyz" for 2 dollars for a year. So, what I did was to redirect the traffic received from this address into the elastic IP so that it then redirects the traffic to the EC2 instance. Also tried to deploy the infrastructure and it worked.... sort of. I still need the DNS records to propagate properly.

Other changes:
- Made the instance accessible to the discord bot that's supposed to automate the starting and stopping of the server.