Minecraft and Terraria Server Provisioning
Added files for provisioning both a minecraft and terraria server in the EC2 instance running on AWS. For this I just copied the notes that I had for setting up a minecraft server and followed this guide (https://terraria.fandom.com/wiki/Server) on setting up a terraria server on Linux. Included some additional files in the repo like the Eula that the user is supposed to agree to for the minecraft server, as well as the server icon and the server.properties.