New Copy Worlds Playbook and More Provisioning Elements
During the provisioning step I found out that the act of copying worlds from the host container to the server took way too long (up to 40 minutes). So I decided to separate the actions for copying both terraria and minecraft worlds into their own playbook. Now the provisioning step could be executed pretty without waiting for so long. I also added a new entrypoint script in "docker-compose.yaml" to execute the worlds playbook.

Aside from this, I also modified the provisioning playbook to now install Java and copy a pair of startup scripts for both the terraria and minecraft servers. Also, since the worlds were copied into a "worlds" folder, I changed the "server.properties" of the minecraft server to search for the default world in there.