Conditionals for "copy_worlds.yaml"
Decided to add a new conditional for the copy worlds playbook. Now you can pass "extra-vars" to the entrypoint command of the "docker-compose.yaml" file for copying worlds, specifying for which game you want to copy the worlds. In this case I left one option with minecraft and another for terraria. Also removed the additional step of installing unzip separately during the Terraria provisioinig and moved it to be installed alongside of tmux.

Other changes:
- Included the world that we are going to use for terraria and changed the start_server.sh script to run that worlds specifically. However, the command doesn't seem to run as expected. Will have to investigate that.