Added a Server Config File (WIP)
The start_server script was failing because I added arguments to it. So I decided to remove the arguments and add the parameters that were there into a server config file for the server itself. I thought it would work, but the server keeps prompting me for settings each time I run it. I will have to properly check how to make the server run with those configs.

Other changes:
- Made it so that the worlds that are copied into the server are able to be written, read and executed for any user. Basically changed their permissions.