Automatic Namecheap Nameserver Addition
Each time I take the architecture of the server up, I have to redirect the namecheap domain to the nameservers of AWS. This is an unnecesary step in my opinion, so I just added a terraform provider for adding the nameservers automatically. However! Apparently namecheap only gives the ability to access the site with an API with an account balance higher than 50$. So for now all the programming will remain commented. So sad :(