Bugfixes and Starting Terraria Server without Interactive Shell
Introduced changes to make it so that the server can start without having to go through the interactive setup that appears when you execute the previous start_server.sh script. To make this work I had to add the "-config serverconfig.txt" parameter to the start_server.sh script. Because of this I also included a new server config with more parameters for when we want to run the server next time.

Other changes:
- Removed the server files for Windows and Mac. We only use the Linux ones anyways
- Ignored some temporary files that appear while you configure your infrastructure with Terraform
- When running the server, the server starts creating backups from time to time of the world. However, the server process did not have permissions to write to the Worlds folder, so each time a backup was created the server crashed. Added an ansible task that gave read and write access to the folder.