Dynamically Fetching my IP
Found out today that my IP changed, causing the spinned up gaming server to not allow me access through SSH or any other port. Made it so that now my IP is fetched dynamically from "ipv4.icanhazip.com" and is then converted into a local variable that can be used across my main.tf script. I love this. The only thing that I have to take into account is that namecheap requires you to whitelist the IPs that should be accessible. If my IP changes, then I have to change the whitelisted IP found in namecheap. That is the only manual step in the process. The server provisioning should still be possible, only the addition of the namecheap nameservers changes.