Added S3 Bucket and IAM Role for Persistent Storage
Decided to add an S3 bucket for things like worlds and other configurations used in the gaming server. However, you simply cannot access S3 data from the EC2 instance, so I also created an IAM role that allows the EC2 instance to access the S3 bucket. To create the role I used this (https://aws.amazon.com/es/blogs/security/writing-iam-policies-how-to-grant-access-to-an-amazon-s3-bucket/) as reference to create the S3 access policy and this (https://skundunotes.com/2021/11/16/attach-iam-role-to-aws-ec2-instance-using-terraform/) to create the role and all the other requirements.

Other changes:
- Parametrized the EC2 instance type. Now I can easily change it from the locals variables of the script.