Parametrized the Terraria Version
Terraria just had a minor update, causing us to not be able to access the server, since the server was on version 1447 and our game was on version 1449. So, I decided to parametrize the Terraria version in the provisioning playbook, so that in the future, in case the game updates again, I can just change the value of the variable and everything should run correctly when doing "./start_server.sh". Pretty cool.