Minecraft Java and File Access Fixes
The "start_server.sh" script for minecraft was not working because of two things: First, there was no Java version that could run the server, and after multiple tries I couldn't get any Java version to install properly. Second, once Java was installed, the server.properties and eula.txt files were not accesible by the server.jar file. The two changes that I had to apply were:
- Installing the java SDK through a regular shell command, instead of through the APT module that comes with Ansible. I'm still not a 100% sure why this works, but it does work... so... I will leave it at that.
- Had to change the permissions of the "minecraft" folder, as well as the files "server.properties", "eula.txt" and the "server-icon.png" to 777. Now the "server.jar" file has the ability to change the configuration files and is also is able to create new files inside of "~/minecraft".

The only thing that I don't like about this is that minecraft doesn't have a good way to download a specific version of the server. One has to go ahead and manually set the link to another value because the links use hashes instead of a comprehensive structure like the one for terraria. There are sites like "mcversions.net" that pair the versions and their hash, but they don't have things like an API that I can query to get the new version. One thing that I could do, is use the Node "mcversion" package to query the versions and get the download link for a specific version... however, this would require me to add a couple of more components to the server, like downloading node, installing the package, running a short script, getting the output of the script and then using that output for the next step of the Ansible playbook. I think its too difficult for such a simple thing. I can just change the link using mcversions.net as a reference.