Terraform S3 Backend
I've had to provision the gaming server multiple times from either my laptop or my desktop. The only problem about doing it from different computers, is that the terraform.tfstate file that is used to track the current state of your deployment has to be sent over to the other computer in order to have the terraform scripts run correctly. In order to not having to do this every time, I configured the terraform S3 backend. Now the terraform.tfstate file is saved in an S3 bucket, which can then be read by terraform from any machine. However... how do we provision the S3 bucket used for the backend? Well I created a new set of terraform scripts inside the terraform folder, that provision a private S3 bucket that is meant to store the terraform state. Now there are two sets of scripts inside of the terraform folder: "aws_server" to actually provision the server and "s3_backend" for configuring the backend that is going to be used by the aws_server.

The only required change that I had to implement for the logic of the original "aws_server" scripts, is the place where the final "hosts.tpl" is rendered. Previously the script only went 1 folder up (since it was just a matter of getting out of the terraform folder). However, now, since there are subfolders inside of the terraform folder, we need to go two levels up.