S3 Backups and Removal of the "Copy Worlds" Playbook
Finally figured out how to backup both the Terraria and Minecraft worlds to an S3 bucket. Instead of using that damn s3fs package (supposedly syncs the filesystem with an S3 bucket), I moved to the AWS CLI. Now, when provisioning the site, all files from an S3 bucket are copied onto the server, and afterwards, a CRON job takes over that periodically syncs the contents of both the Terraria and Minecraft worlds back into the S3 bucket. Its a pretty neat system, although one thing is still lacking: The ability to dynamically fetch new files from the S3 bucket, in order to put them in the server. I would like to be able to add new worlds in the S3 bucket, and then see them appear in the server automatically. Right now, this transfer from S3 to server is only done when running the provisioning script.

Also, because now I don't copy the worlds from my local directory to the server, I removed the "copy_worlds.yaml" and their calls in the docker-compose.yaml file. Now I simply add the cron jobs and do the initial sync on the filesystem through the provisioning script.