Bugfix: Dynamic Change of Terraria Version Failing
Each new terraria version is unzipped onto a differently named folder (eg. version 1.4.4.9 unzips into the folder "1449"). This means that the start_server.sh script must dynamically change the route of the server executable in order to properly run the Terraria server. To do this, I used the ansible "replace" module, which looked for the string "$TERRARIA_VERSION" and replaced it with the version number. This was supposed to work properly, but I found out that the "$TERRARIA_VERSION" string, was actually being interpreted as an environment variable and not giving the replace module the chance to actually change the version. Now, I simply removed the dollar sign at the beggining of the string and the replacement works properly. Great.