Bugfix: No Error on S3 Creation
Terraform follows the mantra of being purely declarative, meaning that it does not concern itself with checking whether a resource exists or not. If a resource already exists, it will crash since its expecting that all the listed resources are meant to be created, not updated. This was the problem that I was facing. When running the terraform script that created all the cloud infrastructure for the server, the script always crashed at the s3 bucket creation section, since the bucket already existed. So I made a quick change: I moved the creation of the S3 bucket for persistent storage to its own folder, and then just included a reference to the resource in the main script through a "data" block. Now, you simply need to run the S3 persistent storage script to have the gaming server script run correctly and without any exceptions. After solving this, all the IAM roles were applied correctly and the AWS CLI in charge of syncing the world saves to the cloud is running correctly

Other changes:
- Renamed the "s3_backend" folder to "s3_terraform_backend"
- Included a new folder with the script to create the "s3_gaming_server_data"
- The S3 bucket for the persistent data, also needs a "terraform.tfstate" file. So, I renamed the tfstate file for the gaming server and added a prefix to the new tfstate file for the persistent storage ("gaming-data-terraform.tfstate").
- Changed the S3 gaming data resource from "main.tf" to a data block that simply contains a "bucket" field.
- Since the process of setting up the server is getting more and more involved, I added more documentation to the Readme file, detailing the different steps that should be taken to properly configure each section.