Removing "Read Only" from id_rsa
Found out that the ":ro" tag at the end of volumes in docker compose is not actually changing the chmod attributes of the file, its simply making it so that the files inside of the folder are not writeable. You are basically removing your ability to change the chmod permissions of your files. So the ":ro" tag was removed, since it actually doesn't do anything. Sadge